# 07-ui-development/12-openapi-clients/mission-ui/mission-service-openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/impostor/{userId}":{"post":{"tags":["game-controller"],"operationId":"guessImpostor","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/accounts/{accountId}/orders":{"post":{"tags":["order-controller"],"operationId":"submitOrder","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequestDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderResponseDto"}}}}},"security":[{"bearerAuth":[]}]}},"/public":{"get":{"tags":["mission-controller"],"operationId":"publicEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/mission":{"get":{"tags":["mission-controller"],"operationId":"missionEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/impostor/word/{userId}":{"get":{"tags":["game-controller"],"operationId":"impostorWord","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/hello":{"get":{"tags":["hello-controller"],"operationId":"hello","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/advisors/{advisorId}":{"get":{"tags":["portfolio-controller"],"operationId":"getAdvisor","parameters":[{"name":"advisorId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Advisor"}}}}}}},"/accounts/{accountId}/transactions":{"get":{"tags":["portfolio-controller"],"operationId":"getTransactions","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Transaction"}}}}}}}}},"components":{"schemas":{"OrderRequestDto":{"type":"object","properties":{"ticker":{"type":"string","minLength":1},"instrumentType":{"type":"string","enum":["EQUITY","BOND","FUND"]},"quantity":{"type":"number","format":"double","exclusiveMinimum":0},"price":{"type":"number","format":"double","exclusiveMinimum":0},"side":{"type":"string","enum":["BUY","SELL"]},"buy":{"type":"boolean"}},"required":["instrumentType","side","ticker"]},"OrderResponseDto":{"type":"object","properties":{"status":{"type":"string"},"fee":{"type":"number","format":"double"},"newHoldingQuantity":{"type":"number","format":"double"}}},"Advisor":{"type":"object","properties":{"advisorId":{"type":"integer","format":"int32"},"name":{"type":"string"},"region":{"type":"string"}}},"Transaction":{"type":"object","properties":{"transactionId":{"type":"integer","format":"int32"},"ticker":{"type":"string"},"txnType":{"type":"string"},"quantity":{"type":"number"},"price":{"type":"number"},"txnDate":{"type":"string","format":"date"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}