# 07-ui-development/12-openapi-clients/mission-ui/src/app/generated/mission-api-client/model/advisor.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Advisor { 
    advisorId?: number;
    name?: string;
    region?: string;
}