# 07-ui-development/12-openapi-clients/mission-ui/src/app/generated/mission-api-client/model/orderRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderRequestDto { 
    ticker: string;
    instrumentType: OrderRequestDto.InstrumentTypeEnum;
    quantity?: number;
    price?: number;
    side: OrderRequestDto.SideEnum;
    buy?: boolean;
}
export namespace OrderRequestDto {
    export const InstrumentTypeEnum = {
        Equity: 'EQUITY',
        Bond: 'BOND',
        Fund: 'FUND'
    } as const;
    export type InstrumentTypeEnum = typeof InstrumentTypeEnum[keyof typeof InstrumentTypeEnum];
    export const SideEnum = {
        Buy: 'BUY',
        Sell: 'SELL'
    } as const;
    export type SideEnum = typeof SideEnum[keyof typeof SideEnum];
}